fix errors

+++ scr/app/task_types/router.py
@@ -13,14 +13,22 @@ from scr.app.task_types.schemas import (
 )
 from scr.app.task_types.dependences import task_type_by_id
 
+from scr.app.core.models import User
+
+from scr.app.auth.router import get_current_user
+
 import scr.app.task_types.crud as crud
 
 router = APIRouter(prefix="/task_type", tags=["Task type"])
 
 
 @router.get("/", response_model=list[TaskType])
-async def get_task_types(session: AsyncSession = Depends(get_async_session)):
-    return await crud.get_task_types(session=session)
+async def get_task_types(
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
+):
+    options = {"user_id": user.id}
+    return await crud.get_task_types(session=session, options=options)
 
 
 @router.post(
@@ -31,12 +39,28 @@ async def get_task_types(session: AsyncSession = Depends(get_async_session)):
 async def create_task_type(
     task_type_in: TaskTypeCreate,
     session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
 ):
-    return await crud.create_task_type(session=session, task_type_in=task_type_in)
+    return await crud.create_task_type(
+        session=session, task_type_in=task_type_in, user_id=user.id
+    )
 
 
 @router.get("/{task_type_id}/", response_model=TaskType)
-async def get_task_type(task_type: TaskType = Depends(task_type_by_id)):
+async def get_task_type(
+    task_type_id: int,
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
+):
+    options = {"user_id": user.id}
+    task_type = await crud.get_task_type(
+        session=session, task_type_id=task_type_id, options=options
+    )
+    if task_type is None:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Task Type not found",
+        )
     return task_type
 
 
@@ -45,7 +69,13 @@ async def update_task_type(
     task_type_update: TaskTypeUpdate,
     task_type: TaskType = Depends(task_type_by_id),
     session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
 ):
+    if task_type.user_id != user.id:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Task Type not yours",
+        )
     return await crud.update_task_type(
         session=session, task_type=task_type, task_type_update=task_type_update
     )
@@ -56,7 +86,13 @@ async def update_task_type_partial(
     task_type_update: TaskTypeUpdatePartial,
     task_type: TaskType = Depends(task_type_by_id),
     session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
 ):
+    if task_type.user_id != user.id:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Task Type not yours",
+        )
     return await crud.update_task_type(
         session=session,
         task_type=task_type,
@@ -69,5 +105,11 @@ async def update_task_type_partial(
 async def delete_task_type(
     task_type: TaskType = Depends(task_type_by_id),
     session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(get_current_user()),
 ) -> None:
+    if task_type.user_id != user.id:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Task Type not yours",
+        )
     await crud.delete_task_type(session=session, task_type=task_type)
